[CST-18016] Improved doc
