+++ dspace-api/src/main/java/org/dspace/importer/external/openalex/metadatamapping/OpenAlexDateMetadataProcessor.java
@@ -17,14 +17,38 @@ import org.apache.logging.log4j.Logger;
 import org.dspace.importer.external.metadatamapping.contributor.AbstractJsonPathMetadataProcessor;
 
 /**
+ * A metadata processor that extracts and validates date values from JSON data.
+ * This processor retrieves a date string from a specified JSON path and ensures
+ * that it is correctly formatted according to the ISO 8601 date standard.
+ *
+ * <p>Example JSON:
+ * <pre>
+ * {
+ *   "published_date": "2023-05-10"
+ * }
+ * </pre>
+ *
+ * If the `path` is set to `"/published_date"`, the extracted value will be `"2023-05-10"`.
+ *
+ * <p>This class extends {@link AbstractJsonPathMetadataProcessor}, which handles JSON parsing and path extraction.
+ *
  * @author Adamo Fapohunda (adamo.fapohunda at 4science.com)
- **/
+ */
 public class OpenAlexDateMetadataProcessor extends AbstractJsonPathMetadataProcessor {
 
     private static final Logger log = LogManager.getLogger(OpenAlexDateMetadataProcessor.class);
 
     private String path;
 
+
+    /**
+     * Extracts and validates a date value from a JSON node.
+     * The extracted value must conform to the ISO 8601 format (YYYY-MM-DD).
+     *
+     * @param node The JSON node containing the date as a text value.
+     * @return The parsed date as a string in ISO 8601 format.
+     * @throws IllegalArgumentException If the node is null, not textual, or does not contain a valid date.
+     */
     @Override
     protected String getStringValue(JsonNode node) {
 
@@ -41,16 +65,31 @@ public class OpenAlexDateMetadataProcessor extends AbstractJsonPathMetadataProce
         }
     }
 
+    /**
+     * Provides the logger instance for error reporting.
+     *
+     * @return The {@link Logger} instance.
+     */
     @Override
     protected Logger getLogger() {
         return log;
     }
 
+    /**
+     * Retrieves the JSON path expression used to extract the date value.
+     *
+     * @return The JSONPath expression as a string.
+     */
     @Override
     protected String getPath() {
         return path;
     }
 
+    /**
+     * Sets the JSON path from which the date value should be extracted.
+     *
+     * @param path The JSONPath expression specifying the location of the date in the JSON structure.
+     */
     public void setPath(String path) {
         this.path = path;
     }
