+++ dspace-api/src/main/java/org/dspace/importer/external/openalex/metadatamapping/OpenAlexPersonFieldMapping.java
@@ -13,6 +13,10 @@ import jakarta.annotation.Resource;
 import org.dspace.importer.external.metadatamapping.AbstractMetadataFieldMapping;
 
 /**
+ * An implementation of {@link AbstractMetadataFieldMapping} responsible for
+ * defining the mapping of the OpenAlex Author metadatum fields on the DSpace
+ * Person metadatum fields
+ *
  * @author Adamo Fapohunda (adamo.fapohunda at 4science.com)
  **/
 public class OpenAlexPersonFieldMapping extends AbstractMetadataFieldMapping {
